Refactor image comparison tests.
